Added validations to account password length
