+++ backend/Controllers/Abstract/ReadOnlyController.cs
@@ -59,7 +59,7 @@ public abstract class ReadOnlyController<TEntity, TDto> : ControllerBase where T
     public virtual async Task<ActionResult<PaginationResult<TDto>>> Query([FromQuery] PaginationQuery query)
     {
         IQueryable<TEntity> queryable = repository.GetAll();
-        var result = await Paginate(queryable, query);
+        var result = await PaginationUtils.Paginate(queryable, query);
 
         if (typeof(TEntity) != typeof(TDto))
         {
@@ -93,56 +93,5 @@ public abstract class ReadOnlyController<TEntity, TDto> : ControllerBase where T
         }
 
         return Ok(result);
-    }
-
-    protected async Task<PaginationResult<TEntity>> Paginate(IQueryable<TEntity> queryable, PaginationQuery query)
-    {
-        int page = Math.Max(1, query.Page.GetValueOrDefault(1));
-        int pageSize = Math.Max(1, query.PageSize.GetValueOrDefault(10));
-
-        if (query.Sort != null)
-        {
-            var sortDir = query.SortDir switch
-            {
-                PaginationQuery.SortAscending => query.SortDir,
-                PaginationQuery.SortDescending => query.SortDir,
-                _ => PaginationQuery.SortAscending
-            };
-
-            var property = ReflectionUtils.GetPropertyIgnoreCase(typeof(TEntity), query.Sort!);
-
-            if (property != null)
-            {
-                if (sortDir == PaginationQuery.SortAscending)
-                {
-                    queryable = queryable.OrderByProperty(property.Name);
-                }
-                else
-                {
-                    queryable = queryable.OrderByDescendingProperty(property.Name);
-                }
-            }
-        }
-
-        var items = await queryable
-            .Skip((page - 1) * pageSize)
-            .Take(pageSize)
-            .ToListAsync();
-
-        var totalPages = (int)Math.Ceiling((double)queryable.Count() / pageSize);
-        var totalCount = await queryable.CountAsync();
-
-        var pageResult = new PaginationResult<TEntity>
-        {
-            Items = items,
-            Page = page,
-            PageSize = Math.Min(items.Count, pageSize),
-            TotalCount = totalCount,
-            TotalPages = totalPages,
-            HasNextPage = page < totalPages,
-            HasPreviousPage = page > 1
-        };
-
-        return pageResult;
     }
 }
